+++ src/Footer.js
@@ -2,7 +2,6 @@ import React                      from 'react';
 import { Link }                   from 'react-router-dom';
 import { ReactComponent as Logo } from './Logo.svg';
 
-
 const Footer = () => {
     
     return(
@@ -10,6 +9,7 @@ const Footer = () => {
             <Logo/>
             <Link to = '/'>Home</Link>
             <Link to = '/changelog'>What's new?</Link>
+            <Link to = '/replace'>Replace</Link>
             <a href = 'https://github.com/erikmartinjordan/jsonmatic'>GitHub</a>
         </div>
     );